# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  domain.ProductAttributes:
    properties:
      description:
        type: string
      expiration_date:
        type: string
      expiration_rate:
        type: number
      freezing_rate:
        type: number
      height:
        type: number
      length:
        type: number
      netweight:
        type: number
      product_code:
        type: string
      product_type_id:
        type: integer
      recommended_freezing_temperature:
        type: number
      seller_id:
        type: integer
      width:
        type: number
    type: object
...
  domain.ProductDoc:
    properties:
      description:
        type: string
      expiration_date:
        type: string
      expiration_rate:
        type: number
      freezing_rate:
        type: number
      height:
        type: number
      id:
        type: integer
      length:
        type: number
      netweight:
        type: number
      product_code:
        type: string
      product_type_id:
        type: integer
      recommended_freezing_temperature:
        type: number
      seller_id:
        type: integer
      width:
        type: number
    type: object
  domain.ProductResponse:
    properties:
      data:
        $ref: '#/definitions/domain.ProductDoc'
    type: object
  domain.ProductsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.ProductDoc'
        type: array
    type: object
...
  domain.ReportProductRecord:
    properties:
      description:
        type: string
      product_id:
        type: integer
      records_count:
        type: integer
    type: object
  domain.ReportResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.ReportProductRecord'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestionar sellers y warehouses
  title: API de Sellers y Warehouses
  version: "1.0"
paths:
...
  /productRecords:
    get:
      description: Devuelve todos los reportes de productos y filtra por id de producto
      parameters:
      - description: ID del producto
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de reportes obtenida exitosamente
          schema:
            $ref: '#/definitions/domain.ReportResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Devuelve todos los reportes de productos y filtra por id de producto
      tags:
      - ProductRecord
  /products:
    get:
      description: Retorna una lista de todos los productos registrados en el sistema
      produces:
      - application/json
      responses:
        "200":
          description: Listado de productos
          schema:
            $ref: '#/definitions/domain.ProductsResponse'
        "400":
          description: Error al parsear la fecha de caducidad
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Productos no encontrados
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Lista todos los productos
      tags:
      - Product
    post:
      description: Crea un nuevo producto en el sistema
      parameters:
      - description: Producto a crear
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.ProductAttributes'
      produces:
      - application/json
      responses:
        "201":
          description: Producto creado exitosamente
          schema:
            $ref: '#/definitions/domain.ProductResponse'
        "409":
          description: Tipo de producto o vendedor no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Crea un nuevo producto
      tags:
      - Product
  /products/{id}:
    delete:
      description: Elimina un producto por su ID
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Producto eliminado exitosamente
          schema:
            $ref: '#/definitions/domain.ProductResponse'
        "400":
          description: ID de producto inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Producto no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Error de eliminación por clave foránea
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Elimina un producto por su ID
      tags:
      - Product
    get:
      description: Obtiene un producto por su ID
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Producto encontrado
          schema:
            $ref: '#/definitions/domain.ProductResponse'
        "400":
          description: ID de producto inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Producto no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Obtiene un producto por su ID
      tags:
      - Product
    patch:
      description: Actualiza un producto por su ID
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      - description: Producto a actualizar
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.ProductAttributes'
      produces:
      - application/json
      responses:
        "200":
          description: Producto actualizado exitosamente
          schema:
            $ref: '#/definitions/domain.ProductResponse'
        "400":
          description: ID de producto inválido
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Producto no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Conflicto de producto
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Actualiza un producto por su ID
      tags:
      - Product
...
swagger: "2.0"